restructured utils to separate plot and logging utils
